pyproject.toml
[tool.poetry]
name = "objectstore-interface"
version = "0.1.0"
description = ""
authors = ["Reilly Hewitson <reilly.hewitson@stfc.ac.uk>"]
readme = "README.md"
packages = [{include = "objectstore_interface"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
authlib = "^1.2.0"
uvicorn = "^0.21.1"
httpx = "^0.23.3"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
requests = "^2.28.2"
python-dateutil = "^2.8.2"
python-multipart = "^0.0.6"
jsonpickle = "^3.0.1"
pytest-asyncio = "^0.21.0"


[tool.poetry.group.dev.dependencies]
bs4 = "^0.0.1"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
